requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.9
django-stdimage==5.1.1
Pillow==7.2.0
progressbar2==3.53.1
psycopg2-binary==2.8.6
python-utils==2.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
